Join common code of Androidstudio and Intellij Plugin installation
